Moved logic from CommandsBot to Listener
